// src/src/core/constants/enums.ts
export enum ActiveInactiveStatus {
  Active = 1,
  Inactive = 0,
}

export enum Status {
  Active = 1,
  Inactive = 0,
  Probation = 2,
  Suspended = 3,
}